Format the golden files using the imports pkg to avoid formatting related test failures
The golden files are formatted using a specific version of Go and
commited.
The generated tests are always formnatted using the current Go version.

The are then comapred against each other.

This bvroke with Go 1.10 due to changes to the formatting rules.

A previous PR fixed this using a *brittle approach*: strip tabs from the
both the goldens and the generated tests before comapring. See #62

@shurcooL propsed a much more elgant solution: after reading the
goldens, format them using the current Go version before comparing.

This PR does just that.